# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Inventory Management API
  version: 1.0.0
  description: API for managing inventory, products, users, and other related entities.
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
paths:
...
  /replenishment/forecast/generate:
    post:
      summary: Trigger demand forecast generation
      description: Triggers a demand forecasting process for a product or all products
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForecastGenerationRequest'
      responses:
        '200':
          description: Forecast generation status
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Demand forecast completed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /replenishment/forecast/{forecastId}:
    parameters:
      - in: path
        name: forecastId
        schema:
          type: integer
        required: true
        description: Forecast ID
    get:
      summary: Get a specific demand forecast
      description: Retrieves details of a specific demand forecast by its ID
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Demand forecast details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemandForecast'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /replenishment/suggestions:
    get:
      summary: Get a list of reorder suggestions
      description: Retrieves a list of suggested reorders based on forecast and stock levels
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, PO_CREATED]
          required: false
          description: Filter by suggestion status
        - in: query
          name: supplierId
          schema:
            type: integer
          required: false
          description: Filter by Supplier ID
      responses:
        '200':
          description: A list of reorder suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReorderSuggestion'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /replenishment/suggestions/{suggestionId}/create-po:
    post:
      summary: Create a draft Purchase Order from a reorder suggestion
      description: Creates a draft Purchase Order based on a selected reorder suggestion
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: suggestionId
          schema:
            type: integer
          required: true
          description: Reorder Suggestion ID
      responses:
        '201':
          description: Purchase Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /replenishment/purchase-orders/{poId}/approve:
    post:
      summary: Approve a draft Purchase Order
      description: Approves a draft Purchase Order, changing its status to APPROVED
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: poId
          schema:
            type: integer
          required: true
          description: Purchase Order ID
      responses:
        '200':
          description: Purchase Order approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /replenishment/purchase-orders/{poId}/send:
    post:
      summary: Send a Purchase Order to the supplier
      description: Marks an approved Purchase Order as SENT
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: poId
          schema:
            type: integer
          required: true
          description: Purchase Order ID
      responses:
        '200':
          description: Purchase Order sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /replenishment/purchase-orders/{poId}:
    parameters:
      - in: path
        name: poId
        schema:
          type: integer
        required: true
        description: Purchase Order ID
    get:
      summary: Get a Purchase Order by ID
      description: Retrieves details of a specific Purchase Order by its ID
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Purchase Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update a Purchase Order
      description: Updates details of a specific Purchase Order
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePORequest'
      responses:
        '200':
          description: Purchase Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /replenishment/purchase-orders/{poId}/receive:
    post:
      summary: Record received goods for a Purchase Order
      description: Records received quantities for items in a Purchase Order and updates stock levels
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: poId
          schema:
            type: integer
          required: true
          description: Purchase Order ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceivePORequest'
      responses:
        '200':
          description: Goods received and stock updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Cannot receive goods for Purchase Order in DRAFT status
        '500':
          $ref: '#/components/responses/InternalServerError'
  /replenishment/purchase-orders/{poId}/cancel:
    post:
      summary: Cancel a Purchase Order
      description: Cancels a Purchase Order if it's in DRAFT or APPROVED status
      tags:
        - Replenishment
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: poId
          schema:
            type: integer
          required: true
          description: Purchase Order ID
      responses:
        '200':
          description: Purchase Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseOrder'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Cannot cancel Purchase Order in SENT status
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ForecastGenerationRequest:
      type: object
      required:
        - periodInDays
      properties:
        periodInDays:
          type: integer
          format: int32
          minimum: 1
          example: 30
        productId:
          type: integer
          format: uint
          nullable: true
          example: 1
    DemandForecast:
      type: object
      properties:
        ID:
          type: integer
          format: int64
          readOnly: true
          example: 1
        CreatedAt:
          type: string
          format: date-time
          readOnly: true
        UpdatedAt:
          type: string
          format: date-time
          readOnly: true
        DeletedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        ProductID:
          type: integer
          format: uint
          example: 1
        ForecastPeriod:
          type: string
          example: 30_DAYS
        PredictedDemand:
          type: integer
          example: 150
        GeneratedAt:
          type: string
          format: date-time
    ReorderSuggestion:
      type: object
      properties:
        ID:
          type: integer
          format: int64
          readOnly: true
          example: 1
        CreatedAt:
          type: string
          format: date-time
          readOnly: true
        UpdatedAt:
          type: string
          format: date-time
          readOnly: true
        DeletedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        ProductID:
          type: integer
          format: uint
          example: 1
        SupplierID:
          type: integer
          format: uint
          example: 1
        CurrentStock:
          type: integer
          example: 50
        PredictedDemand:
          type: integer
          example: 120
        SuggestedOrderQuantity:
          type: integer
          example: 70
        LeadTimeDays:
          type: integer
          example: 7
        Status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, PO_CREATED]
          example: PENDING
        SuggestedAt:
          type: string
          format: date-time
    PurchaseOrder:
      type: object
      properties:
        ID:
          type: integer
          format: int64
          readOnly: true
          example: 1
        CreatedAt:
          type: string
          format: date-time
          readOnly: true
        UpdatedAt:
          type: string
          format: date-time
          readOnly: true
        DeletedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        SupplierID:
          type: integer
          format: uint
          example: 1
        Status:
          type: string
          enum: [DRAFT, APPROVED, SENT, PARTIALLY_RECEIVED, RECEIVED, CANCELLED]
          example: DRAFT
        OrderDate:
          type: string
          format: date-time
        ExpectedDeliveryDate:
          type: string
          format: date-time
          nullable: true
        ActualDeliveryDate:
          type: string
          format: date-time
          nullable: true
        CreatedBy:
          type: integer
          format: uint
          example: 1
        ApprovedBy:
          type: integer
          format: uint
          nullable: true
          example: 1
        ApprovedAt:
          type: string
          format: date-time
          nullable: true
        PurchaseOrderItems:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseOrderItem'
    PurchaseOrderItem:
      type: object
      properties:
        ID:
          type: integer
          format: int64
          readOnly: true
          example: 1
        CreatedAt:
          type: string
          format: date-time
          readOnly: true
        UpdatedAt:
          type: string
          format: date-time
          readOnly: true
        DeletedAt:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        PurchaseOrderID:
          type: integer
          format: uint
          example: 1
        ProductID:
          type: integer
          format: uint
          example: 1
        OrderedQuantity:
          type: integer
          example: 100
        ReceivedQuantity:
          type: integer
          example: 50
        UnitPrice:
          type: number
          format: float
          example: 10.50
    UpdatePORequest:
      type: object
      properties:
        supplierId:
          type: integer
          format: uint
          example: 2
        status:
          type: string
          enum: [DRAFT, APPROVED, SENT, RECEIVED, CANCELLED]
          example: APPROVED
        orderDate:
          type: string
          format: date-time
          example: "2023-10-27T10:00:00Z"
        expectedDeliveryDate:
          type: string
          format: date-time
          example: "2023-11-15T10:00:00Z"
        items:
          type: array
          items:
            $ref: '#/components/schemas/POItemRequest'
    POItemRequest:
      type: object
      required:
        - productId
        - orderedQuantity
        - unitPrice
      properties:
        productId:
          type: integer
          format: uint
          example: 1
        orderedQuantity:
          type: integer
          format: int32
          minimum: 1
          example: 50
        unitPrice:
          type: number
          format: float
          minimum: 0
          example: 12.00
    ReceivePORequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReceivePOItemRequest'
...